LayoutTests/scrollingcoordinator/mac/latching/horizontal-overflow-back-swipe-expected.txt

Scrolls vertically
Scrolls horizontally
A back swipe, over a scrolled-to-left overflow, should start a navigation gesture.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Swipe to the left
PASS minXOffset is 0
PASS maxXOffset is -1000
PASS successfullyParsed is true

TEST COMPLETE
